refactor: normalize Heading nodes
