+++ packages/rich-text/src/helpers/__tests__/extractNodes.test.ts
@@ -2,8 +2,8 @@ import { createEditor as createSlateEditor } from '@udecode/plate-test-utils';
 import { PlateEditor } from '@udecode/plate-core';
 import { Path } from 'slate';
 import { BLOCKS, INLINES } from '@contentful/rich-text-types';
-import { extractParagraphsAt } from './editor';
-import { CustomElement } from '../types';
+import { extractParagraphs } from '../extractNodes';
+import { CustomElement } from '../../types';
 
 const createEditor = (children: CustomElement[]) =>
   createSlateEditor('test-editor', {}, children) as PlateEditor;
@@ -74,7 +74,7 @@ const buildList = (
   ],
 });
 
-describe('slateNodeEntryToText', () => {
+describe('extractParagraphs', () => {
   it('table', () => {
     const table: CustomElement = {
       data: {},
@@ -124,7 +124,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([table]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([
+    expect(extractParagraphs(editor, path)).toEqual([
       paragraph('header 1'),
       paragraph('header 2'),
       paragraph('cell 1'),
@@ -142,7 +142,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([blockquote]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([paragraph('text 1'), paragraph('text 2')]);
+    expect(extractParagraphs(editor, path)).toEqual([paragraph('text 1'), paragraph('text 2')]);
   });
 
   it('list - UL', () => {
@@ -150,7 +150,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([ul]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([
+    expect(extractParagraphs(editor, path)).toEqual([
       paragraph('text 1'),
       paragraph('text 2'),
       paragraph('text 3'),
@@ -162,7 +162,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([ol]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([
+    expect(extractParagraphs(editor, path)).toEqual([
       paragraph('text 1'),
       paragraph('text 2'),
       paragraph('text 3'),
@@ -182,7 +182,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([element]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([
+    expect(extractParagraphs(editor, path)).toEqual([
       paragraph('text 1', { bold: true, italic: true }),
       paragraph('text 2', { underline: true, code: true }),
     ]);
@@ -207,7 +207,7 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([element]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([paragraphWithLink, paragraph('text 2')]);
+    expect(extractParagraphs(editor, path)).toEqual([paragraphWithLink, paragraph('text 2')]);
   });
 
   it('should preserve embedded inline entries', () => {
@@ -229,6 +229,6 @@ describe('slateNodeEntryToText', () => {
     const editor = createEditor([element]);
     const path: Path = [0];
 
-    expect(extractParagraphsAt(editor, path)).toEqual([paragraphWithEmbedded, paragraph('text 2')]);
+    expect(extractParagraphs(editor, path)).toEqual([paragraphWithEmbedded, paragraph('text 2')]);
   });
 });
