+++ packages/rich-text/src/plugins/Hyperlink/EntryAssetTooltip.tsx
@@ -5,7 +5,7 @@ import {
 } from '@contentful/field-editor-reference/dist/types';
 import { entityHelpers, Entry, Asset, FieldExtensionSDK } from '@contentful/field-editor-shared';
 import { getScheduleTooltipContent } from '@contentful/field-editor-reference';
-import { truncate } from '../../helpers/truncate';
+import { truncate } from './truncate';
 
 function getEntityInfo({
   entityTitle,
